+++ lib/commands/index.d.ts
@@ -1,5 +1,5 @@
 import { Command, CommandEvent, Result } from "./../types";
-export declare const commands: (CommandEvent: CommandEvent) => Promise<Result | {
+export declare const commands: (CommandEvent: CommandEvent) => Promise<void | Result | {
     qoyod_total: number;
     repzo_total: number;
     disabled: number;